move MetaInfo to a seperate file and move metadata generation into Metainfo
