+++ srcs/andhra.py
@@ -7,6 +7,7 @@ import urllib
 from ..utils import utils
 from .basegazette import BaseGazette
 from .central import CentralBase
+from ..utils.metainfo import MetaInfo
 
 class Andhra(BaseGazette):
     def __init__(self, name, storage):
@@ -109,7 +110,7 @@ class Andhra(BaseGazette):
         return minfos
 
     def parse_row(self, tr, order, dateobj):
-        metainfo = utils.MetaInfo()
+        metainfo = MetaInfo()
         metainfo.set_date(dateobj)
 
         i = 0
@@ -263,7 +264,7 @@ class AndhraBase(BaseGazette):
         return minfos
 
     def parse_row(self, tr, order, dateobj):
-        metainfo = utils.MetaInfo()
+        metainfo = MetaInfo()
         metainfo.set_date(dateobj)
 
         i = 0
@@ -446,7 +447,7 @@ class AndhraArchive(CentralBase):
         if not download:
             return
  
-        metainfo = utils.MetaInfo()
+        metainfo = MetaInfo()
         metainfos.append(metainfo)
         metainfo.set_date(dateobj)
         metainfo['download'] = download
